Fix wrong type

+++ c/prolog_list.h
@@ -1,3 +1,3 @@
 #include <SWI-Prolog.h>
 
-int calculate_pl_list_len(term_t pl_list);
+size_t calculate_pl_list_len(term_t pl_list);
